+++ tests/auxiliary/json.rs
@@ -146,7 +146,7 @@ mod tests {
         .unwrap()
         .filter_map(Result::ok)
         .collect();
-        assert_eq!(files.len(), 28);
+        assert_eq!(files.len(), 38);
 
         for file in files {
             let s = fs::read_to_string(file).unwrap();